+++ NEWS.md
@@ -4,6 +4,14 @@ points for further improvements.
 
 
 
+### Version in 3.6.3
+Date: 23.07.2026 - progress
+
+Checks on number of groups in Taylor modelling added. Now actively advises
+on 't' parameter reduction.
+
+
+
 ### Version in 3.6.2
 Date: 20.07.2026 - release